refactor: change after review
