+++ src/features/layout/layout.module.scss
@@ -1,5 +1,5 @@
 .authLayout {
   display: flex;
+  align-items: center;
   justify-content: center;
-  padding-top: 30px;
 }
