+++ src/features/layout/layout.tsx
@@ -2,27 +2,27 @@ import { PropsWithChildren, ReactNode } from 'react'
 
 import { Header } from '@photo-fiesta/ui-lib'
 
-import style from './authLayout.module.scss'
+import style from './layout.module.scss'
 
 /**
- * `AuthLayout` is a layout component that wraps around pages that require authentication.
+ * `Layout` is a layout component that wraps around pages that require authentication.
  * It includes a header with user-specific options and provides a consistent layout
  * for authentication-related pages.
  *
  * @component
  * @param {PropsWithChildren} props - The props object containing child components.
- * @returns {JSX.Element} The `AuthLayout` component containing a header and the child content.
+ * @returns {JSX.Element} The `Layout` component containing a header and the child content.
  *
  * @example
  * return (
- *   <AuthLayout>
+ *   <Layout>
  *     <SignInPage />
- *   </AuthLayout>
+ *   </Layout>
  * )
  */
 
-const AuthLayout = ({ children }: PropsWithChildren) => {
-  //isLoggedIn & onChangeLanguage is mock data
+const Layout = ({ children }: PropsWithChildren) => {
+  //toDO isLoggedIn & onChangeLanguage
   const isLoggedIn = true
   const onChangeLanguage = (value: string) => {
     console.log(`Language changed to ${value}`)
@@ -37,15 +37,15 @@ const AuthLayout = ({ children }: PropsWithChildren) => {
 }
 
 /**
- * `getAuthLayout` is a helper function that applies the `AuthLayout` to a given page.
+ * `getAuthLayout` is a helper function that applies the `Layout` to a given page.
  *
- * @param {ReactNode} page - The page component to be wrapped by the `AuthLayout`.
- * @returns {JSX.Element} The page component wrapped in the `AuthLayout`.
+ * @param {ReactNode} page - The page component to be wrapped by the `Layout`.
+ * @returns {JSX.Element} The page component wrapped in the `Layout`.
  *
  * @example
  * SignUpPage.getLayout = getAuthLayout
  */
 
 export const getAuthLayout = (page: ReactNode) => {
-  return <AuthLayout>{page}</AuthLayout>
+  return <Layout>{page}</Layout>
 }
